% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deepRsq.R
\name{calc_deepRsq}
\alias{calc_deepRsq}
\title{Calculate deepRsq values using pre-trained deepRsq models}
\usage{
calc_deepRsq(
  file,
  header = T,
  model,
  FeatureCols = 5:85,
  keptCols = 1:6,
  outfile = NULL,
  seed = 123
)
}
\arguments{
\item{file}{Required. A vector of strings with each element being one file name. Could be a single file or multiple files (if separated by chromosomes).}

\item{header}{Logical. The parameter is passing to the \code{fread} function. TRUE by default. Strongly recommended containing a header in your data file.}

\item{model}{Required. The pre-trained deepRsq model used as references for calculating deepRsq values. Output from the \code{train_deepRsq} function.}

\item{FeatureCols}{Vector of column numbers of all the features used for model training. By default, it will be 5:85 including 81 features covering
population genetics, allele frequencies across populations, etc., provided online. More user-defined features are also feasible.
Note that the default values are different from \code{train_deepRsq} function since we assume true R2 information is not available in the input data here.}

\item{keptCols}{Vector of column numbers that needs to be carried over into the output files, e.g. variants ID, chromosome, positions, alleles, estimated MAF. etc.
Default values 1:6, including Chr, Pos, Ref, Alt, MAF and Rsq.}

\item{outfile}{Name of the output file containing the calculated deepRsq values for each variant. Strongly recommended to specify the output file name to save the data.}

\item{seed}{Random seed. Default value 123.}
}
\value{
Dataframe of the calculated deepRsq values, along with the information carried over from the input data (keptCols of the input data).
}
\description{
This function is used to calculate deepRsq values given pre-trained deepRsq models from \code{train_deepRsq} function.
Variant level features in the input data should be exactly the same as the input data for \code{train_deepRsq} function.
}
\examples{

MESA_deepRsq = calc_deepRsq(file = "MESA_1000G_allchr.txt.gz", 
model = JHS_model, FeatureCols = 6:86, keptCols = 1:7, 
outfile = "MESA_deepRsq_JHS_model.txt.gz")

## If data are separated by chromosomes 
WHI_deepRsq = calc_deepRsq(file = paste0("WHI_1000G_chr",1:22,".txt.gz"), 
model = JHS_model,  outfile = "WHI_deepRsq_JHS_model.txt.gz")

}
