% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MagicalRsq.R
\name{calc_MagicalRsq}
\alias{calc_MagicalRsq}
\title{Calculate MagicalRsq values using pre-trained MagicalRsq models}
\usage{
calc_MagicalRsq(
  file,
  header = T,
  model,
  FeatureCols = 5:83,
  keptCols = 1:6,
  MAF_cate = NULL,
  MAFCol = NULL,
  outfile = NULL,
  seed = 123
)
}
\arguments{
\item{file}{Required. A vector of strings with each element being one file name. Could be a single file or multiple files (if separated by chromosomes).}

\item{header}{Logical. The parameter is passing to the \code{fread} function. TRUE by default. Strongly recommended containing a header in your data file.}

\item{model}{Required. The pre-trained MagicalRsq model used as references for calculating MagicalRsq values. Output from the \code{train_MagicalRsq} function.}

\item{FeatureCols}{Vector of column numbers of all the features used for model training. By default, it will be 5:83 including 79 features covering
population genetics, allele frequencies across populations, etc., provided online. More user-defined features are also feasible.
Note that the default values are different from \code{train_MagicalRsq} function since we assume true R2 information is not available in the input data here.}

\item{keptCols}{Vector of column numbers that needs to be carried over into the output files, e.g. variants ID, chromosome, positions, alleles, estimated MAF. etc.
Default values 1:6, including Chr, Pos, Ref, Alt, MAF and Rsq.}

\item{MAF_cate}{Minor allee frequency category. Must be chosen from c("common","lowfreq","rare). If not specified, it will use all the variants to train models.}

\item{MAFCol}{Column number of minor allele frequency. Optional. Must be specified if \code{MAF_cate} is used.}

\item{outfile}{Name of the output file containing the calculated MagicalRsq values for each variant. Strongly recommended to specify the output file name to save the data.}

\item{seed}{Random seed. Default value 123.}
}
\value{
Dataframe of the calculated MagicalRsq values, along with the information carried over from the input data (keptCols of the input data).
}
\description{
This function is used to calculate MagicalRsq values given pre-trained MagicalRsq models from \code{train_MagicalRsq} function.
Variant level features in the input data should be exactly the same as the input data for \code{train_MagicalRsq} function.
}
\examples{

toy_MagicalRsq = calc_MagicalRsq(file = "test_chr22.txt.gz", 
model = toy_model, FeatureCols = 6:84, keptCols = 1:7, 
outfile = "test_MagicalRsq.txt.gz")

## If data are separated by chromosomes 
toy2_MagicalRsq = calc_MagicalRsq(file = paste0("toy2_chr",1:22,".txt.gz"), 
model = toy2,  outfile = "toy2_MagicalRsq.txt.gz")

}
