% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_deepRsq.R
\name{eval_deepRsq}
\alias{eval_deepRsq}
\title{Evaluate deepRsq performance (compared to true R2)}
\usage{
eval_deepRsq(
  file,
  header = T,
  MAC = NULL,
  N = NULL,
  MAF = NULL,
  trueR2Col = 5,
  MAFCol = 6,
  RsqCol = 7,
  deepRsqCol = 8
)
}
\arguments{
\item{file}{Required. A vector of strings with each element being one file name. Could be a single file or multiple files (if separated by chromosomes).}

\item{header}{Logical. The parameter is passing to the \code{fread} function. TRUE by default. Strongly recommended containing a header in your data file.}

\item{MAC}{Disgard variants with extremely low minor allele counts for the evaluation purpose. Must be used together with \code{N} and should not be used together with \code{MAF}.}

\item{N}{Total sample size in terms of individuals in the dataset. Used to calculate MAF as the variants inclusion criteria. Must be used together with \code{MAC} should not be used together with \code{MAF}.}

\item{MAF}{Disgard variants with extremely low minor allele frequencies for the evaluation purpose. Should not be used together with \code{MAC} or \code{N}.
Note that if \code{MAC},\code{N}, and \code{MAF} are all provided, \code{MAF} will be used as the primary threshold, and the other two parameters won't take any effects.}

\item{trueR2Col}{Column number of true R2. Default value 5.}

\item{MAFCol}{Column number of estimated minor allele frequency. Default value 6.}

\item{RsqCol}{Column number of Rsq. Default value 7.}

\item{deepRsqCol}{Column number of deepRsq. Default value 8.}
}
\value{
A named one-row dataframe with three metrics: R2 (squared Pearson correlation with true R2), RMSE (root mean square error) and MAE (mean absolute error) for both Rsq and deepRsq. Six columns in total.
}
\description{
This function is used to evaluate deepRsq values calculated by \code{calc_deepRsq} in the presence of true R2.
}
\examples{

MESA_eval = eval_deepRsq(file = "MESA_deepRsq_JHS_model.txt.gz", MAC = 5, N = 2000)

BioMe_eval = eval_deepRsq(file = paste0("BioMe_deepRsq_MESA_model_chr",1:22,".txt.gz"), 
MAF = 0.001, trueR2Col = 8, MAFCol = 5, RsqCol = 6, deepRsqCol = 7)

}
