% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_MagicalRsq.R
\name{eval_MagicalRsq}
\alias{eval_MagicalRsq}
\title{Evaluate MagicalRsq performance (compared to true R2)}
\usage{
eval_MagicalRsq(
  file,
  header = T,
  minMAC = NULL,
  N = NULL,
  minMAF = NULL,
  trueR2Col = 5,
  MAFCol = 6,
  RsqCol = 7,
  MagicalRsqCol = 8
)
}
\arguments{
\item{file}{Required. A vector of strings with each element being one file name. Could be a single file or multiple files (if separated by chromosomes).}

\item{header}{Logical. The parameter is passing to the \code{fread} function. TRUE by default. Strongly recommended containing a header in your data file.}

\item{minMAC}{Disgard variants with extremely low minor allele counts for the evaluation purpose. Must be used together with \code{N} and should not be used together with \code{MAF}.}

\item{N}{Total sample size in terms of individuals in the dataset. Used to calculate MAF as the variants inclusion criteria. Must be used together with \code{MAC} should not be used together with \code{MAF}.}

\item{minMAF}{Disgard variants with extremely low minor allele frequencies for the evaluation purpose. If specified some cutoff c, it will only evaluate variants with \code{MAF > c}. Should not be used together with \code{minMAC} or \code{N}.
Note that if \code{minMAC},\code{N}, and \code{minMAF} are all provided, \code{minMAF} will be used as the primary threshold, and the other two parameters won't take any effects.}

\item{trueR2Col}{Column number of true R2. Default value 5.}

\item{MAFCol}{Column number of estimated minor allele frequency. Default value 6.}

\item{RsqCol}{Column number of Rsq. Default value 7.}

\item{MagicalRsqCol}{Column number of MagicalRsq. Default value 8.}
}
\value{
A named one-row dataframe with three metrics: R2 (squared Pearson correlation with true R2), RMSE (root mean square error) and MAE (mean absolute error) for both Rsq and MagicalRsq. Six columns in total.
}
\description{
This function is used to evaluate MagicalRsq values calculated by \code{calc_MagicalRsq} in the presence of true R2.
}
\examples{

test_eval = eval_MagicalRsq(file = "test_MagicalRsq.txt.gz", minMAC = 5, N = 2000)

test2_eval = eval_MagicalRsq(file = paste0("test2_chr",1:22,".txt.gz"), 
minMAF = 0.001, trueR2Col = 8, MAFCol = 5, RsqCol = 6, MagicalRsqCol = 7)

}
